tests: Move to new file
