+++ tests/Strings/StringsTest.php
@@ -172,128 +172,4 @@ class StringsTest extends TestCase
     {
         $this->assertSame($expected, str($value1)->slice($value2, $value3));
     }
-
-    /**
-     * @test
-     * @dataProvider noCaseProvider
-     */
-    public function noCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::noCase($actual));
-    }
-
-    /** @test */
-    public function noCaseWithOpt()
-    {
-        $this->assertSame('camel#case', Str::noCase('camelCase', '#'));
-        $this->assertSame('camel#case', Str::noCase('#camel Case#', '#'));
-    }
-
-    /**
-     * @test
-     * @dataProvider camelCaseProvider
-     */
-    public function camelCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::camelCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider capitalCaseProvider
-     */
-    public function capitalCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::capitalCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider constantCaseProvider
-     */
-    public function constantCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::constantCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider dotCaseProvider
-     */
-    public function dotCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::dotCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider headerCaseProvider
-     */
-    public function headerCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::headerCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider pascalCaseProvider
-     */
-    public function pascalCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::pascalCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider pathCaseProvider
-     */
-    public function pathCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::pathCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider sentenceCaseProvider
-     */
-    public function sentenceCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::sentenceCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider snakeCaseProvider
-     */
-    public function snakeCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::snakeCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider spinalCaseProvider
-     */
-    public function spinalCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::spinalCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider swapCaseProvider
-     */
-    public function swapCase($expected, $actual)
-    {
-        $this->assertSame($expected, Str::swapCase($actual));
-    }
-
-    /**
-     * @test
-     * @dataProvider titleCaseProvider
-     */
-    public function titleCase($actual, $expected)
-    {
-        $this->assertSame($expected, Str::titleCase($actual));
-    }
 }
